requirements.txt
google-adk>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
